Compile DIAL client and server for macOS.
Changes:
* Added method to get mac and ip address for OSX. Current Linux method
of query via SIOCGIFCONF doesn't work. Added code to read via getifaddrs() API.
* Put non OSX specific code under #ifdef flags.
* Removed few linker libraries which are not required (lcares and lrt).
* Fixed many warnings.

Tested on: OSX 10.12

+++ client/DialServer.cpp
@@ -226,7 +226,7 @@ int DialServer::getStatus(
 
     if (!status) return 0;
     ATRACE("Body: %s\n", responseBody.c_str());
-    unsigned found = responseBody.find("href=");
+    size_t found = responseBody.find("href=");
     if( found != string::npos )
     {
         // The start of href is after href= and the quote
